Convert to SDK-style project
The library targets ".NET Standard 2.0", to be usable in .NET Framework
and .NET Core projects. The tool chain is .NET Core 3.1 instead of the
current .NET-5 because it's a long-term-support release. The next LTS,
.NET-6, is not yet released.
The project compiles, the tests can be executed, but some of them fail.
At a first glance, reading is the issue. To be investigated.