Work around NetFx vs. NetCore differences
The .NET Framework impplementation of System.IO.Packaging, in
WindowsBase.dll, does call into GetPartCore in many more places then the
.NET Core implementation from the System.IO.Packaing nuget, see GetPartHelper:
- https://referencesource.microsoft.com/#WindowsBase/Base/System/IO/Packaging/Package.cs
- https://github.com/dotnet/runtime/blob/main/src/libraries/System.IO.Packaging/src/System/IO/Packaging/Package.cs

We now work around this by calling out GetParts from the ctor to populate
the cache. (Note that the Package.Open implementation does this for
ZipPackage, too).

Unfortunately, this was not enough as the our GetPartsCore implementation
was broken. It was already broken in the .NET Framework version, but not
discovered by the tests. Now this is checked, too, as part  of
`AssertPackagesEqual`.