Adaprt GitVersion.yml to play nice with strong-naming and SemVer
To be able to replace a dependency without recompilation of the
consumer, the `AssemblyVersion` must not change (otherwise, binding
redicects are required). Following the rules of Semantic Versioning,
there are no breaking changes within the same major. Thus we only
change/increment the Assembly Version in case of major-bumb.

For Windows Installer to replace a file on disk during an upgrade, the
`AssemblyFileVersion` has to be larger. Thus we have to ensure newer
versions have a larger value here. We cannot naively use the pre-release
number here as this one is reseted then changing the pre-release label.
E.g. after "alpha.3" comes "beta.1" -- non the less the beta has to
replace the alpha. For this to happen, we put a "weight" on the
prerelease, as the file version only allows integers. The weight is
choosen to still for in 16 bit *and* make it easy for humans to tell
them apart.

Last but not least, the mainline increment is changed from "patch" to
"minor" as I expect that the master branch heads for new feature
releases by default. For occasional hotfixes, only incrementing patch, a
dedicated branch may be used.